requirements.txt
amqp==5.1.1
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.5
django-dotenv==1.4.2
dnspython==2.4.2
eventlet==0.33.3
greenlet==2.0.2
gunicorn==21.2.0
idna==3.4
kombu==5.3.2
MarkupSafe==2.1.3
mock==5.1.0
nameko==2.14.1
packaging==23.1
path==16.7.1
path.py==12.5.0
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.0.5
vine==5.0.0
Werkzeug==2.3.7
wrapt==1.15.0
